postgres-fastapi/pyproject.toml
[project]
name = "postgres-fastapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.37",
    "uvicorn[standard]>=0.42.0",
]
